pyproject.toml
[project]
name = "fresh-alert-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "langchain>=1.0.0a10",
    "langchain-anthropic>=0.3.21",
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langfuse>=3.5.2",
    "langgraph>=1.0.0a4",
    "langgraph-cli[inmem]>=0.4.2",
    "pillow>=11.3.0",
]
